requirements.txt
# django
django==3.2.15
django-cors-headers==3.2.1
djangorestframework==3.11.2
django-filter==2.4.0
pandas==1.0.1
cython==0.29.23
pyproj==3.0.1
geopandas==0.8.0
Fiona~=1.8.19
numpy==1.18.1
Pillow==8.3.2
django-storages==1.9.1
boto3==1.12.1
geopy==1.21.0
Shapely==1.7.0
unidecode==1.2.0
drf-yasg==1.20.0
django-cte==1.2.1

# web server and auth
oauth2client==4.1.3
requests==2.22.0

# postgresql and sql
psycopg2-binary==2.8.5
django-ltree==0.5.3

# webpack loader
# This is our custom version of webpack loader to handle S3 better.
# Use the following line if you want a locally editable version, and comment the other
# -e git+https://github.com/BLSQ/django-webpack-loader@e94f76d0f8372193f0b662e2a1aba01a9deffb20#egg=django_webpack_loader
# Use the tar.gz version to install faster
https://github.com/BLSQ/django-webpack-loader/archive/e94f76d0f8372193f0b662e2a1aba01a9deffb20.tar.gz#egg=django_webpack_loader

# importing tools
jsonschema==3.2.0


# exporting tools
xlsxwriter==1.2.7
beautifulsoup4==4.7.1
lxml==4.9.1
djangorestframework-csv==2.1.1
openpyxl==3.0.10

# dhis2 api wrapper
dhis2.py==2.0.2

# odk-specific libraries
pyxform==1.1.0


ipython==7.12.0
django-extensions==2.2.8
lxml==4.9.1
xlrd==1.2.0
pyjwt==1.7.1
djangorestframework_simplejwt==4.4.0

sentry-sdk==1.5.11
django-lazy-services==0.0.3
dateparser==1.0.0

gspread==5.0.0
gspread-formatting==1.0.5
oauth2client==4.1.3

django-contrib-comments~=2.1.0
django-sql-dashboard~=1.1

# Allauth for wfp
django-allauth==0.50.0


django-json-widget==1.1.1

# Seems to be required to run the project
django-stubs==1.9.0
